fix: coins value styling
