+++ packages/zignaly-ui/src/components/filters/ZigFilters/dropdowns/MultiFilterDropdown/styles.ts
@@ -1,5 +1,4 @@
 import { styled } from "@mui/system";
-import ZigTypography from "components/display/ZigTypography";
 import { DropdownItem } from "../../styles";
 
 export const LayoutItem = styled(DropdownItem)`
@@ -7,17 +6,4 @@ export const LayoutItem = styled(DropdownItem)`
   align-items: center;
   justify-content: center;
   gap: 11px;
-`;
-
-export const FiltersCount = styled(ZigTypography)`
-  background: ${({ theme }) => theme.palette.neutral500};
-  width: 16px;
-  height: 16px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: ${({ theme }) => theme.palette.links};
-  font-weight: 600;
-  font-size: 12px;
-  border-radius: 50%;
 `;
